Move views into the actual views folder
