+++ app/public/components/flags/flagsIndexChip.js
