+++ app/public/components/sidebar/sidebar.js
@@ -13,7 +13,7 @@
  */
 
 import { h } from '/js/src/index.js';
-import { RCT } from '../../../config.js';
+import { RCT } from '../../config.js';
 import sidebarItem from './sidebarItem.js';
 const { pageNames } = RCT;
