+++ app/public/components/sidebar/sidebarItem.js
