+++ app/public/components/sidebar/subPagesCount.js
