+++ app/public/views/flags/detectorName.js
