+++ app/public/views/flags/flagVisualization.js
