+++ app/public/views/flags/flagsTable.js
@@ -16,7 +16,7 @@ import { h } from '/js/src/index.js';
 import { RCT } from '../../../../config.js';
 import flagsTableHeader from './table/flagsTableHeader.js';
 import flagsTableRow from './table/flagsTableRow.js';
-import pagesCellsSpecials from '../pagesCellsSpecials.js';
+import pagesCellsSpecials from '../userView/data/pagesCellsSpecials.js';
 const { pageNames: PN } = RCT;
 
 export default function flagsTable(model, run, detector) {
