+++ app/public/views/flags/flagsVisualization.js
