+++ app/public/views/flags/table/fields.js
