+++ app/public/views/flags/table/flagsTableRow.js
