+++ app/public/views/userView/data/pagesCellsSpecials.js
@@ -14,7 +14,7 @@
 import { h } from '/js/src/index.js';
 import { RCT } from '../../../config.js';
 import { getReadableFileSizeString } from '../../../utils/utils.js';
-import actionChip from '../../common/actionChip.js';
+import actionChip from '../../../components/common/actionChip.js';
 const { dataReqParams: DRP } = RCT;
 const { pageNames: PN } = RCT;
