+++ app/public/views/userView/data/posting/postingDataConfig.js
