+++ app/public/views/userView/data/table/filtering/activeFilters.js
