+++ app/public/views/userView/data/table/indexChip.js
