+++ app/public/views/userView/data/table/items-counter.js
