+++ app/public/views/userView/data/table/noDataView.js
