+++ app/public/views/userView/data/table/noSubPageSelected.js
