+++ app/public/views/userView/data/table/tablePanel.js
